small refactor in speech problems
