+++ tensor2tensor/data_generators/speech_recognition.py
@@ -19,14 +19,11 @@ it as appropriate (e.g. using apt-get or yum).
 """
 
 import functools
-import os
-from subprocess import call
-import tempfile
 import numpy as np
-from scipy.io import wavfile
 import scipy.signal
 
 from tensor2tensor.data_generators import problem
+from tensor2tensor.data_generators import audio_encoder
 from tensor2tensor.data_generators import text_encoder
 from tensor2tensor.layers import common_attention
 from tensor2tensor.layers import common_layers
@@ -153,83 +150,6 @@ def compute_mel_filterbank_features(
   return tf.expand_dims(log_mel_sgram, -1, name="mel_sgrams")
 
 
-#
-# Audio problem definition
-#
-class AudioEncoder(object):
-  """Encoder class for saving and loading waveforms."""
-
-  def __init__(self, num_reserved_ids=0, sample_rate=16000):
-    assert num_reserved_ids == 0
-    self._sample_rate = sample_rate
-
-  @property
-  def num_reserved_ids(self):
-    return 0
-
-  def encode(self, s):
-    """Transform a string with a filename into a list of float32.
-
-    Args:
-      s: path to the file with a waveform.
-
-    Returns:
-      samples: list of int16s
-    """
-    # Make sure that the data is a single channel, 16bit, 16kHz wave.
-    # TODO(chorowski): the directory may not be writable, this should fallback
-    # to a temp path, and provide instructions for installing sox.
-    if s.endswith(".mp3"):
-      # TODO(dliebling) On Linux, check if libsox-fmt-mp3 is installed.
-      out_filepath = s[:-4] + ".wav"
-      call([
-          "sox", "--guard", s, "-r", "16k", "-b", "16", "-c", "1", out_filepath
-      ])
-      s = out_filepath
-    elif not s.endswith(".wav"):
-      out_filepath = s + ".wav"
-      if not os.path.exists(out_filepath):
-        call(["sox", "-r", "16k", "-b", "16", "-c", "1", s, out_filepath])
-      s = out_filepath
-    rate, data = wavfile.read(s)
-    assert rate == self._sample_rate
-    assert len(data.shape) == 1
-    if data.dtype not in [np.float32, np.float64]:
-      data = data.astype(np.float32) / np.iinfo(data.dtype).max
-    return data.tolist()
-
-  def decode(self, ids):
-    """Transform a sequence of float32 into a waveform.
-
-    Args:
-      ids: list of integers to be converted.
-
-    Returns:
-      Path to the temporary file where the waveform was saved.
-
-    Raises:
-      ValueError: if the ids are not of the appropriate size.
-    """
-    _, tmp_file_path = tempfile.mkstemp()
-    wavfile.write(tmp_file_path, self._sample_rate, np.asarray(ids))
-    return tmp_file_path
-
-  def decode_list(self, ids):
-    """Transform a sequence of int ids into an image file.
-
-    Args:
-      ids: list of integers to be converted.
-
-    Returns:
-      Singleton list: path to the temporary file where the wavfile was saved.
-    """
-    return [self.decode(ids)]
-
-  @property
-  def vocab_size(self):
-    return 256
-
-
 class ByteTextEncoderWithEos(text_encoder.ByteTextEncoder):
   """Encodes each byte to an id and appends the EOS token."""
 
@@ -279,7 +199,7 @@ class SpeechRecognitionProblem(problem.Problem):
         "inputs": None,  # Put None to make sure that the logic in
                          # decoding.py doesn't try to convert the floats
                          # into text...
-        "waveforms": AudioEncoder(),
+        "waveforms": audio_encoder.AudioEncoder(),
         "targets": ByteTextEncoderWithEos(),
     }
